EnvirronmentalReadingService: Split into Interface and an implementation
